Touch-ups and modernization
The most significant change is that I changed the wording to not be totally hostile towards uncompressed texture data (it's still both useful and common, and the only option that guarantees 100% fidelity).
Further, I added a hint that PNG is better than JPEG if you choose to use standard file formats rather than DXT-compressed DDS files.
Last, the AMD Compressonator is deprecated and no longer supported, so I changed the link to point to the modern tool. An image showing the GUI of the deprecated tool was lost in the process. Consider adding that illustration back for the new tool.